requirements.txt
asgiref==3.8.1
Django==5.2.1
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Deprecated==1.2.18
dj-database-url==2.3.0
Django==5.2.1
django-environ==0.12.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-secret-manager==2.23.3
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
opentelemetry-api==1.32.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.30.2
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==0.19.0
requests==2.32.3
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
wrapt==1.17.2
zipp==3.21.0
asgiref==3.8.1
Django==5.2.1
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
